% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{create_viz}
\alias{create_viz}
\title{Create Visualization}
\usage{
create_viz(
  df,
  horizontal = FALSE,
  vertical = FALSE,
  round = 2,
  lab_col = c("#da654f", "#d78f54", "#8aba8f", "#4b6b91", "#7eacc5")
)
}
\arguments{
\item{df}{the dataframe you wish to use}

\item{horizontal}{takes a TRUE or FALSE argument for whether you want horizontal spacing in your sankey chart}

\item{vertical}{takes a TRUE or FALSE argument for whether you want vertical spacing in your sankey chart}

\item{round}{takes a numeric argument for what value you want to round your labels to}

\item{lab_col}{take a vector for the colors you want for your risk factors (in alphabetical order of your risk factors)}
}
\value{
a ggplot of your sankey chart
}
\description{
This function takes a dataframe as an input. It assumes that the data in the
columns contains the risk values over time, with years as the column names and
risk factors as the row names.
}
