% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{process_data}
\alias{process_data}
\title{Process data}
\usage{
process_data(df)
}
\arguments{
\item{df}{the dataframe you wish to use}
}
\value{
a new dataframe in the format to create the sankey later on
}
\description{
This function takes a dataframe as an input. It assumes that the data in the
columns contains the risk values over time, with years as the column names and
risk factors as the row names.
}
